Keep cleanign up code with passing tests
