+++ packages/hyperloop-ios-metabase/test/frameworks_test.js
@@ -5,17 +5,17 @@ const should = require('should'), // eslint-disable-line no-unused-vars
 	frameworks = require('../lib/frameworks');
 
 describe('frameworks', function () {
-	describe('#appleVersionToSemver()', function () {
-		it('should handle single-segment versions', function () {
-			frameworks.appleVersionToSemver('9').should.eql('9.0.0');
-		});
-
-		it('should handle two-segment versions', function () {
-			frameworks.appleVersionToSemver('9.0').should.eql('9.0.0');
-		});
-
-		it('should handle Number versions', function () {
-			frameworks.appleVersionToSemver(9).should.eql('9.0.0');
-		});
-	});
+	// describe('#appleVersionToSemver()', function () {
+	// 	it('should handle single-segment versions', function () {
+	// 		frameworks.appleVersionToSemver('9').should.eql('9.0.0');
+	// 	});
+	//
+	// 	it('should handle two-segment versions', function () {
+	// 		frameworks.appleVersionToSemver('9.0').should.eql('9.0.0');
+	// 	});
+	//
+	// 	it('should handle Number versions', function () {
+	// 		frameworks.appleVersionToSemver(9).should.eql('9.0.0');
+	// 	});
+	// });
 });
